Add README
There was an idea, many know this, called TabbedOut. The idea was to bring together a group of remarkable people, to see if they could become something more. See if they could work together when they needed to find great places to eat and drink while testing a mobile payments application.